// src/api/types/repo_state.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RepoStateKind } from "./repo_state_kind";

export type RepoState = { kind: RepoStateKind, step: number | null, total_steps: number | null, head_name: string | null, onto: string | null, message: string, };